+++ lib/generators/shopify_app/home_controller/templates/unauthenticated_home_controller.rb
@@ -3,8 +3,10 @@
 class HomeController < ApplicationController
   include ShopifyApp::EmbeddedApp
   include ShopifyApp::RequireKnownShop
+  include ShopifyApp::ShopAccessScopesVerification
 
   def index
     @shop_origin = current_shopify_domain
+    @host = params[:host]
   end
 end