+++ lib/generators/shopify_app/user_model/templates/user.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 class User < ActiveRecord::Base
-  include ShopifyApp::UserSessionStorage
+  include ShopifyApp::UserSessionStorageWithScopes
 
   def api_version
     ShopifyApp.configuration.api_version