+++ test/generators/shop_model_generator_test.rb
@@ -15,7 +15,7 @@ class ShopModelGeneratorTest < Rails::Generators::TestCase
     run_generator
     assert_file "app/models/shop.rb" do |shop|
       assert_match "class Shop < ActiveRecord::Base", shop
-      assert_match "include ShopifyApp::ShopSessionStorage", shop
+      assert_match "include ShopifyApp::ShopSessionStorageWithScopes", shop
       assert_match(/def api_version\n\s*ShopifyApp\.configuration\.api_version\n\s*end/, shop)
     end
   end
@@ -24,6 +24,24 @@ class ShopModelGeneratorTest < Rails::Generators::TestCase
     run_generator
     assert_migration "db/migrate/create_shops.rb" do |migration|
       assert_match "create_table :shops  do |t|", migration
+    end
+  end
+
+  test "create shop with access_scopes migration in a test environment" do
+    run_generator
+    assert_migration "db/migrate/add_shop_access_scopes_column.rb" do |migration|
+      assert_match "add_column :shops, :access_scopes, :string", migration
+    end
+  end
+
+  test "create shop with access_scopes migration with --new-shopify-cli-app flag provided" do
+    Rails.env = 'mock_environment'
+
+    run_generator %w(--new-shopify-cli-app)
+    Rails.env = 'test' # Change this back for subsequent tests
+
+    assert_migration "db/migrate/add_shop_access_scopes_column.rb" do |migration|
+      assert_match "add_column :shops, :access_scopes, :string", migration
     end
   end