+++ test/generators/user_model_generator_test.rb
@@ -15,7 +15,7 @@ class UserModelGeneratorTest < Rails::Generators::TestCase
     run_generator
     assert_file "app/models/user.rb" do |user|
       assert_match "class User < ActiveRecord::Base", user
-      assert_match "include ShopifyApp::UserSessionStorage", user
+      assert_match "include ShopifyApp::UserSessionStorageWithScopes", user
       assert_match(/def api_version\n\s*ShopifyApp\.configuration\.api_version\n\s*end/, user)
     end
   end
@@ -24,6 +24,24 @@ class UserModelGeneratorTest < Rails::Generators::TestCase
     run_generator
     assert_migration "db/migrate/create_users.rb" do |migration|
       assert_match "create_table :users do |t|", migration
+    end
+  end
+
+  test "create access_scopes migration for User model" do
+    run_generator
+    assert_migration "db/migrate/add_user_access_scopes_column.rb" do |migration|
+      assert_match "add_column :users, :access_scopes, :string", migration
+    end
+  end
+
+  test "create User with access_scopes migration with --new-shopify-cli-app flag provided" do
+    Rails.env = 'mock_environment'
+
+    run_generator %w(--new-shopify-cli-app)
+    Rails.env = 'test' # Change this back for subsequent tests
+
+    assert_migration "db/migrate/add_user_access_scopes_column.rb" do |migration|
+      assert_match "add_column :users, :access_scopes, :string", migration
     end
   end