+++ test/shopify_app/session/session_repository_test.rb
@@ -45,47 +45,40 @@ module ShopifyApp
     test '.retrieve_user_session_by_shopify_user_id retrieves a user session by JWT' do
       SessionRepository.user_storage = InMemoryUserSessionStore
 
-      user_id = 'abra'
-      InMemoryUserSessionStore.expects(:retrieve_by_shopify_user_id).with(user_id)
+      InMemoryUserSessionStore.expects(:retrieve_by_shopify_user_id).with(mock_shopify_user_id)
 
-      SessionRepository.retrieve_user_session_by_shopify_user_id(user_id)
+      SessionRepository.retrieve_user_session_by_shopify_user_id(mock_shopify_user_id)
     end
 
     test '.retrieve_shop_session_by_shopify_domain retrieves a shop session by JWT' do
       SessionRepository.shop_storage = InMemoryShopSessionStore
 
-      shopify_domain = 'abra-shop'
-      InMemoryShopSessionStore.expects(:retrieve_by_shopify_domain).with(shopify_domain)
+      InMemoryShopSessionStore.expects(:retrieve_by_shopify_domain).with(mock_shopify_domain)
 
-      SessionRepository.retrieve_shop_session_by_shopify_domain(shopify_domain)
+      SessionRepository.retrieve_shop_session_by_shopify_domain(mock_shopify_domain)
     end
 
     test '.retrieve_user_session retrieves a user session' do
       SessionRepository.user_storage = InMemoryUserSessionStore
 
-      user_id = 'abra'
-      InMemoryUserSessionStore.expects(:retrieve).with(user_id)
+      InMemoryUserSessionStore.expects(:retrieve).with(mock_shopify_user_id)
 
-      SessionRepository.retrieve_user_session(user_id)
+      SessionRepository.retrieve_user_session(mock_shopify_user_id)
     end
 
     test '.retrieve_shop_session retrieves a shop session' do
       SessionRepository.shop_storage = InMemoryShopSessionStore
 
-      shop_id = 'abra-shop'
-      InMemoryShopSessionStore.expects(:retrieve).with(shop_id)
+      mock_shop_id = 'abra-shop'
+      InMemoryShopSessionStore.expects(:retrieve).with(mock_shop_id)
 
-      SessionRepository.retrieve_shop_session(shop_id)
+      SessionRepository.retrieve_shop_session(mock_shop_id)
     end
 
     test '.store_shop_session stores a shop session' do
       SessionRepository.shop_storage = InMemoryShopSessionStore
 
-      session = ShopifyAPI::Session.new(
-        domain: 'shop.myshopify.com',
-        token: 'abracadabra',
-        api_version: :unstable
-      )
+      session = mock_shopify_session
       InMemoryShopSessionStore.expects(:store).with(session)
 
       SessionRepository.store_shop_session(session)
@@ -94,15 +87,30 @@ module ShopifyApp
     test '.store_user_session stores a user session' do
       SessionRepository.user_storage = InMemoryUserSessionStore
 
-      session = ShopifyAPI::Session.new(
-        domain: 'shop.myshopify.com',
-        token: 'abracadabra',
-        api_version: :unstable
-      )
+      session = mock_shopify_session
       user = { 'id' => 'abra' }
       InMemoryUserSessionStore.expects(:store).with(session, user)
 
       SessionRepository.store_user_session(session, user)
+    end
+
+    private
+
+    def mock_shopify_domain
+      'shop.myshopify.com'
+    end
+
+    def mock_shopify_user_id
+      'abra'
+    end
+
+    def mock_shopify_session
+      ShopifyAPI::Session.new(
+        domain: mock_shopify_domain,
+        token: 'abracadabra',
+        api_version: :unstable,
+        access_scopes: 'read_products'
+      )
     end
   end
 end